+++ src/main/java/foodfinder/repository/GradeRepository.java
@@ -8,15 +8,17 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import javax.transaction.Transactional;
+import java.util.List;
 
 
 @Repository
 public interface GradeRepository extends JpaRepository<RestaurantGrade, Integer> {
 
-    @Modifying(clearAutomatically = true)
     @Transactional
     @Query(value = "SELECT AVG(rg.grade) FROM restaurants_grade rg WHERE rg.id_restaurants = :restaurantId", nativeQuery = true)
     Double countRestaurantGrade(@Param("restaurantId") Integer restaurantId);
+
+    List<RestaurantGrade> findRestaurantGradesByIdRestaurants(Integer idRestaurants);
 
 
 }
