+++ src/main/java/foodfinder/services/interfaces/RestaurantService.java
@@ -1,6 +1,7 @@
 package foodfinder.services.interfaces;
 
 
+import foodfinder.dto.FavoriteRestaurant;
 import foodfinder.dto.Restaurant;
 
 import java.util.List;
@@ -10,4 +11,6 @@ public interface RestaurantService {
     List<Restaurant> fetchRestaurantByNameAndType(final String restaurantName, final List<String> restaurantTypeList);
 
     Restaurant fetchRestaurantId(Integer restaurantId);
+
+    List<Restaurant> fetchFavoriteRestaurantByUserId(Integer userId);
 }
